requirements.txt
fastapi
uvicorn[standard]
requests
pypdf2
python-dotenv
google-generativeai